Add social links to header, remove footer
